Move searchtree input
