+++ src/core/fem/src/geometry/4C_fem_geometry_searchtree_input.cpp
@@ -5,27 +5,26 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "4C_inpar_searchtree.hpp"
+#include "4C_fem_geometry_searchtree_input.hpp"
 
 #include "4C_io_input_spec_builders.hpp"
+
 FOUR_C_NAMESPACE_OPEN
 
-
-
-Core::IO::InputSpec Inpar::Geo::valid_parameters()
+Core::IO::InputSpec Geo::valid_parameters()
 {
   using namespace Core::IO::InputSpecBuilders;
   Core::IO::InputSpec spec = group("SEARCH TREE",
       {
 
-          deprecated_selection<Inpar::Geo::TreeType>("TREE_TYPE",
+          deprecated_selection<Geo::TreeType>("TREE_TYPE",
               {
-                  {"notree", Inpar::Geo::Notree},
-                  {"octree3d", Inpar::Geo::Octree3D},
-                  {"quadtree3d", Inpar::Geo::Quadtree3D},
-                  {"quadtree2d", Inpar::Geo::Quadtree2D},
+                  {"notree", Geo::Notree},
+                  {"octree3d", Geo::Octree3D},
+                  {"quadtree3d", Geo::Quadtree3D},
+                  {"quadtree2d", Geo::Quadtree2D},
               },
-              {.description = "set tree type", .default_value = Inpar::Geo::Notree})},
+              {.description = "set tree type", .default_value = Geo::Notree})},
       {.required = false});
   return spec;
 }
